% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastNoNo-package.R
\docType{package}
\name{fastNoNo-package}
\alias{fastNoNo}
\alias{fastNoNo-package}
\title{fastNoNo: Fast hierarchial Bayesian Normal-Normal models}
\description{
Fits Bayesian hierarchial linear regression models using numerical linear algebra and low dimensional Gaussian quadrature.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pgree/fastNoNo}
  \item Report bugs at \url{https://github.com/pgree/fastNoNo/issues/}
}

}
\author{
\strong{Maintainer}: Philip Greengard \email{philip.greengard@gmail.com}

Authors:
\itemize{
  \item Jonah Gabry \email{jsg2201@columbia.edu}
}

}
\keyword{internal}
