% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mixed_formula.R
\name{fit_mixed_formula}
\alias{fit_mixed_formula}
\title{Formula interface to \code{fit_mixed()}}
\usage{
fit_mixed_formula(
  formula,
  data,
  ...,
  sd_sigma_y,
  sd_sigma1,
  sd_beta2,
  nnt = 20
)
}
\arguments{
\item{formula}{(formula) An \code{lmer}-style model formula. Currently only a
single varying intercept or varying slope is supported in the "random
effects" part of the model formula. To bypass this restriction use
\code{\link[=fit_mixed]{fit_mixed()}} directly.}

\item{data}{(data frame) The data containing the variables used in the model.}

\item{...}{Currently for internal use only.}

\item{sd_sigma_y}{(positive real) Scale parameter value for the prior on
\eqn{\sigma_y}.}

\item{sd_sigma1}{(positive real) Scale parameter value for the prior on
\eqn{\sigma1}.}

\item{sd_beta2}{(positive reals) Scale parameter values for the priors on
\eqn{\beta2} ("fixed effects" coefficients). Must have either one element
or \code{ncol(X2)} elements. In the former case the value is recycled.}

\item{nnt}{(positive integer) Number of quadrature nodes in \eqn{\theta}
direction as described in Greengard et al. (2022). We recommend increasing
\code{nnt} to improve the accuracy of the estimates if the errors are too large
(see \code{errors} slot in returned fitted model object).}
}
\value{
A named list with the following components:
\itemize{
\item \code{beta1}: A data frame with two columns (\code{mean}, \code{sd}) containing the
posterior means and standard deviations for the vector \eqn{\beta1} (the
"random effects").
\item \code{beta2}: A data frame with two columns (\code{mean}, \code{sd})  containing the
posterior means and standard deviations for the vector \eqn{\beta2} (the
"fixed effects").
\item \code{sigma}: A data frame with two columns (\code{mean}, \code{sd}) containing the
posterior means and standard deviations for \eqn{\sigma_y} (the residual
standard deviation) and \eqn{\sigma1} (the standard deviation of
\eqn{\beta1}).
\item \code{cov}: The posterior covariance matrix of \eqn{[\beta1, \beta2]}.
\item \code{errors}: A data frame with two columns (\code{error_mean}, \code{error_sd})
containing the approximate accuracy of the posterior mean and standard
deviation estimates.
\item \code{time}: The execution time of the algorithm in seconds. This only includes
the time taken to run the internal C++ code. To include the full time elapsed
when running the \R function see \code{\link[=system.time]{system.time()}}.
}
}
\description{
This function provides a limited \code{lmer}-style formula interface to the model
fit by \code{\link[=fit_mixed]{fit_mixed()}}. Currently only a small subset of possible formula
specifications are supported.
}
\examples{
fit <- fit_mixed_formula(
  formula = mpg ~ wt + as.factor(gear) + (1|cyl),
  data = mtcars,
  sd_beta2 = 10,
  sd_sigma_y = 10,
  sd_sigma1 = 5
)
fit$beta1
fit$beta2
fit$sigma

# check accuracy of estimates
fit$errors

# refitting using more quadrature nodes improves accuracy
fit <- fit_mixed_formula(
  formula = mpg ~ wt + as.factor(gear) + (1 | cyl),
  data = mtcars,
  sd_beta2 = 10,
  sd_sigma_y = 10,
  sd_sigma1 = 5,
  nnt = 30
)
fit$errors

}
\references{
Philip Greengard, Jeremy Hoskins, Charles C. Margossian, Jonah Gabry, Andrew
Gelman, and Aki Vehtari. (2022). Fast methods for posterior inference of
two-group normal-normal models. To appear,
\href{http://www.stat.columbia.edu/~gelman/research/published/two_group_fastnono.pdf}{Bayesian Analysis}
}
\seealso{
\code{\link[=fit_mixed]{fit_mixed()}}
}
