% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_two_group.R
\name{fit_two_group}
\alias{fit_two_group}
\title{Fast two-group normal-normal model}
\usage{
fit_two_group(y, X1, X2, sd_y, sd_1, sd_2, nnt = 10)
}
\arguments{
\item{y}{Outcome vector.}

\item{X1}{Data matrix corresponding to group 1.}

\item{X2}{Data matrix corresponding to group 2.}

\item{sd_y}{Prior on residual standard deviation.}

\item{sd_1}{Hyperprior on standard deviation of group 1.}

\item{sd_2}{Hyperprior on standard deviation of group 2.}

\item{nnt}{Number of quadrature nodes in \eqn{\theta}. See Greengard et al.
(2021) for details.}
}
\value{
A named list with the following components:
\itemize{
\item \code{beta_1}: A data frame of posterior means and standard deviations for
the vector \eqn{\beta_1}, the coefficients on \code{X1}.
\item \code{beta_2}: A data frame of posterior means and standard deviations for
the vector \eqn{\beta_2}, the coefficients on \code{X2}.
\item \code{sigma}: A data frame with posterior means and standard deviations for
\eqn{\sigma_y}, \eqn{\sigma_1}, and \eqn{\sigma_2}.
\item \code{errors}: A data frame with approximate accuracy of the posterior
mean and standard deviation estimates.
}
}
\description{
This function fits a fast approximation to the Bayesian two-group
hierarchical linear regression model

\deqn{y ~ normal(X_1 \beta_1 + X_2 \beta_2, \sigma_y)}
\deqn{\beta_1 ~ normal(0, \sigma_1)}
\deqn{\beta_2 ~ normal(0, \sigma_2)}
\deqn{\sigma_y ~ normal+(0, sd_y)}
\deqn{\sigma_1 ~ normal+(0, sd_1)}
\deqn{\sigma_2 ~ normal+(0, sd_2)}

The algorithm for computing the fit uses numerical linear algebra and
low dimensional Gaussian quadrature. See Greengard et al. (2021) for details.
}
\examples{
\dontrun{
# Simulate data
n <- 10000
k_1 <- 50
k_2 <- 60

sigma_y <- 1
sigma_1 <- 0.5
sigma_2 <- 2
beta_1 <- rnorm(k_1, 0, sigma_1)
beta_2 <- rnorm(k_2, 0, sigma_2)

X_1 <- matrix(rnorm(n * k_1, 2, 3), ncol = k_1)
X_2 <- matrix(rnorm(n * k_2, -1, 5), ncol = k_2)
y <- rnorm(n, X_1 \%*\% beta_1 + X_2 \%*\% beta_2, sigma_y)

# hyperpriors on scale parameters
sd_1 <- 1
sd_2 <- 2
sd_y <- 2

# Fit model
fit <- fit_two_group(y, X_1, X_2, sd_y, sd_1, sd_2)
str(fit)

# Plot estimates of the betas vs "truth"
plot(fit$beta_1$mean, beta_1); abline(0, 1, col = "red")
plot(fit$beta_2$mean, beta_2); abline(0, 1, col = "red")
}

}
\references{
Philip Greengard, Jeremy Hoskins, Charles C. Margossian, Andrew Gelman, and Aki Vehtari.
(2021). Fast methods for posterior inference of two-group normal-normal models.
\href{https://arxiv.org/abs/2110.03055}{preprint arXiv:2110.03055}
}
