% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_two_group_dense.R
\name{fit_two_group_dense}
\alias{fit_two_group_dense}
\title{Two-group normal-normal model fit}
\usage{
fit_two_group_dense(X1, X2, y)
}
\arguments{
\item{X1}{Data matrix corresponding to group 1.}

\item{X2}{Data matrix corresponding to group 2.}

\item{y}{Outcome vector.}
}
\value{
A list containing posterior means, standard deviations,
and the approximate accuracy of the standard deviations and means. The `mean`
and `std` vectors in the list are of length k+3. The first k entries
correspond to posterior means and standard deviations of the k parameters
\eqn{\beta_1, \beta_2} and the final three correspond to the scale parameters
\eqn{\sigma_1, \sigma_2, \sigma_y}.
}
\description{
This function fits the Bayesian two-group linear regression model
\deqn{y ~ normal(X_1 \beta_1 + X_2 \beta_2, \sigma_y)}
\deqn{\beta_1 ~ normal(0, \sigma_1)}
\deqn{\beta_2 ~ normal(0, \sigma_2)}
\deqn{\sigma_1 ~ normal+(0, 1)}
\deqn{\sigma_2 ~ normal+(0, 1)}
\deqn{\sigma_y ~ normal+(0, 1)}
}
