% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastNoNo-package.R
\docType{package}
\name{fastNoNo-package}
\alias{fastNoNo}
\alias{fastNoNo-package}
\title{fastNoNo: What the package does (short line)}
\description{
More about what it does (maybe more than one line)
}
\keyword{internal}
