% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_two_group_mixed_formula.R
\name{fit_mixed_formula}
\alias{fit_mixed_formula}
\title{Formula interface to \code{fit_mixed()}}
\usage{
fit_mixed_formula(formula, data, ...)
}
\arguments{
\item{formula}{An \code{lmer}-style model formula. Currently only a single varying
intercept or varying slope is supported in the "random effects" part of the
model. To bypass this restriction use \code{\link[=fit_mixed]{fit_mixed()}} directly.}

\item{data}{A data frame containing the variables used in the model.}

\item{...}{Arguments passed to \code{\link[=fit_mixed]{fit_mixed()}}, except for \code{y}, \code{X1},
and \code{X2}, which are generated automatically.}
}
\value{
See \code{\link[=fit_mixed]{fit_mixed()}}.
}
\description{
This function provides a formula interface to the model fit by
\code{\link[=fit_mixed]{fit_mixed()}}.
}
\examples{
fit <- fit_mixed_formula(
  formula = mpg ~ wt + as.factor(gear) + (1|cyl),
  data = mtcars,
  sd_beta2 = 10,
  sd_sigma_y = 10,
  sd_sigma1 = 5
)
fit$beta1
fit$beta2

}
