% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit a two-group normal-normal model}
\usage{
fit_model(X, k1, k2, y)
}
\arguments{
\item{X}{The matrix of data of size n x (k1+k2)}

\item{k1}{size of first group}

\item{k2}{size of second group}

\item{y}{outcome}
}
\value{
fit a list containing posterior means, standard devations,
and accuracy
}
\description{
Fit a two-group normal-normal model
}
