% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_two_group_dense.R
\name{fit_two_group_dense}
\alias{fit_two_group_dense}
\title{Two-group normal-normal model fit}
\usage{
fit_two_group_dense(X1, X2, y, compile_dir = tempdir())
}
\arguments{
\item{X1}{Data matrix corresponding to group 1.}

\item{X2}{Data matrix corresponding to group 2.}

\item{y}{Outcome vector.}

\item{compile_dir}{Path to a directory to store the compiled executable. The
default is \code{\link[=tempdir]{tempdir()}}, in which case compilation will happen once per R
session.}
}
\value{
A named list with the following components:
\itemize{
\item \code{beta_1}: A data frame of posterior means and standard deviations for
the vector \eqn{\beta_1}, the coefficients on \code{X1}.
\item \code{beta_2}: A data frame of posterior means and standard deviations for
the vector \eqn{\beta_2}, the coefficients on \code{X2}.
\item \code{sigma}: A data frame with posterior means and standard deviations for
\eqn{\sigma_y}, \eqn{\sigma_1}, and \eqn{\sigma_2}.
\item \code{error}: A scalar. The approximate accuracy of the posterior mean and
standard deviation estimates.
}
}
\description{
This function fits a fast approximation to the
Bayesian two-group hierarchical linear regression model

\deqn{y ~ normal(X_1 \beta_1 + X_2 \beta_2, \sigma_y)}
\deqn{\beta_1 ~ normal(0, \sigma_1)}
\deqn{\beta_2 ~ normal(0, \sigma_2)}
\deqn{\sigma_1 ~ normal+(0, 1)}
\deqn{\sigma_2 ~ normal+(0, 1)}
\deqn{\sigma_y ~ normal+(0, 1)}

using numerical linear algebra and low dimensional Gaussian quadrature.

Currently the algorithm is implemented in Fortran and using the package
requires the ability to compile Fortran code. Eventually the Fortran code may
come pre-compiled.
}
\examples{
\dontrun{
k1 <- 50
k2 <- 60
n <- 10000

y <- rnorm(n)
X1 <- matrix(rnorm(n * k1, 2, 3), ncol = k1)
X2 <- matrix(rnorm(n * k2, -1, 5), ncol = k2)

fit <- fit_two_group_dense(X1, X2, y)
str(fit)
}

}
