% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_two_group_mixed.R
\name{fit_two_group_mixed_formula}
\alias{fit_two_group_mixed_formula}
\title{Formula interface to \code{fit_two_group_mixed()}}
\usage{
fit_two_group_mixed_formula(data, fixed_formula, varying_intercept, ...)
}
\arguments{
\item{data}{(data frame) The data containing the variables used in the model.}

\item{fixed_formula}{(formula) A formula in the style of \code{\link[stats:lm]{stats::lm()}} that
specifies the outcome variable on the left side and the "fixed effects"
terms on the right side. This formula is used to construct the \code{X2} matrix
for \code{\link[=fit_two_group_mixed]{fit_two_group_mixed()}}.}

\item{varying_intercept}{(string) The name of the grouping variable by which
the intercept should vary (with partial pooling). Currently only a single
varying intercept is supported via this argument. This variable is used to
construct the \code{X1} matrix for \code{\link[=fit_two_group_mixed]{fit_two_group_mixed()}}.}

\item{...}{
  Arguments passed on to \code{\link[=fit_two_group_mixed]{fit_two_group_mixed}}
  \describe{
    \item{\code{y}}{(vector) The outcome variable.}
    \item{\code{X1}}{(matrix) The design matrix for the "random effects" part of the
model. Must have \code{length(y)} rows.}
    \item{\code{X2}}{(matrix) The design matrix for the "fixed effects" part of the
model. Must have \code{length(y)} rows.}
    \item{\code{ss}}{(positive reals) Scale parameter values for the prior on
\eqn{\beta_2}. Must have either one element or \code{ncol(X2)} elements. In the
former case the value is recycled.}
    \item{\code{sd_y}}{(positive real) Scale parameter value for the prior on \eqn{\sigma_y}.}
    \item{\code{sd1}}{(positive real) Scale parameter value for the prior on \eqn{\sigma_1}.}
    \item{\code{nnt}}{(positive integer) Number of quadrature nodes in \eqn{\theta}
direction as described in Greengard et al. (2022).}
  }}
}
\description{
This function fits the same model as \code{\link[=fit_two_group_mixed]{fit_two_group_mixed()}} but the
"fixed effects" part of the model can be specified using a formula.
}
